server/requirements.txt
asgiref==3.8.1
black==25.1.0
boto3==1.38.1
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
Django==5.1.6
django-ckeditor==6.7.0
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
google-auth==2.27.0
gunicorn==23.0.0
idna==3.10
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
Pillow==11.2.1
platformdirs==4.3.7
psycopg==3.2.4
psycopg-binary==3.2.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
